web/requirements.txt
PyJWT>=2.8
cryptography>=42.0
